filesystem:clock_pro: add tests of wal_reader
* Main bug: need to find the offset into each file regardless
  of wether the file is a completely padded mem_aligned or not

  In addition, we must take into account the offset between pages

* Fixed wal_writer not to pad the file
  we just have to deal w/ the fact that the server
  will crash and need to fix half padded files

* Add read_exactly<>() patterns inspired by the
  input_stream_reader<> class.

* use aligned_buffers w/ pmem_aligned*

* tests: add tests for wal_reading using clock-pro

* change types from temporary_buffer<> to
  unique_ptr<char_type[], free_deleter> to use alligned
  buffers

* Add checksums for header and payload

* Add checks to make sure we don't read more than 1 page.

* bugfix: skip bad checksum records.

* we must ONLY skip bad records, but not bad files or requests.
  just because we skip one record, it doesn't mean that the next records
  aren't healthy.

* bugfix: test_runner.py: chdir before running tests

* bugfix: increase timeout to 10s on test_runner.py

+++ src/integration_tests/wal/main.cc
@@ -61,8 +61,10 @@ int main(int args, char **argv, char **env) {
                 return x.append(gen_payload(kPayload)).then([&x](uint64_t i) {
                   auto &p = smf::priority_manager::thread_local_instance()
                               .default_priority();
-                  smf::wal_read_request rq{i, (uint64_t)std::strlen(kPayload),
-                                           uint32_t(0), p};
+                  smf::wal_read_request rq{
+                    static_cast<int64_t>(i),
+                    static_cast<int64_t>(std::strlen(kPayload)), uint32_t(0),
+                    p};
                   return x.get(std::move(rq)).then([i](auto read_result) {
                     assert(read_result);
                     auto size = read_result->fragments.front().data.size();
